Change the manner in which calls to the standalone renderer send some of the key problem settings. Send an encrypted problemJWT instead of sending all data as simple POST data. This is intended to protect the server, and depends on setting the "problemJWTsecret" (shared symmetric encryption key) and the "aud" setting put into the JWT.
"problemJWTsecret" should at present be a 32 character string using
7-bit ASCII characters (so it will behave in both locations it is set).

Both settings must match what is set in the render.conf file for the
Standalone renderer. The "aud" setting must be the same as the value
of SITE_HOST in render.conf.

Note: For testing use with a local Docker standalone renderer, the "aud"
and SITE_HOST value should remain 'http://localhost:3000' which will allow
a local browser to download the htdocs files needed to render a problem
sent by the devstack.