Units where the display_name or problem_banner_text still had the default value could not be copied, as copying does not work when the value of an XBlock field is still a string waiting for "lazy" translation.
The stacktrace would report:
bson.errors.InvalidDocument: cannot encode object: 'WeBWorK Problem', of type: <class 'django.utils.functional.lazy.<locals>.__proxy__'>

The same sort of issue was reported and fixed for the SXA Xblock in
https://github.com/mitodl/edx-sga/issues/252
and
https://github.com/mitodl/edx-sga/pull/254
by changing back to having _ be ugettext instead of ugettext_lazy.

However, for the help strings, and possibly for other things, ugettext_lazy
seems more appropriate.

Thus, we are fixing the bug by using non-lazy ugettext only for where _ was
used to request translation of a field default value.

After this patch copy works, so it is sufficient.