Protect critical Standalone renderer API settings by adding them into the JWT. The Standalone renderer used settings from inside the JWT to override values set by regular form parameters, so this approach prevents tampering with the relevant values.
Sample test code (changing the value of "outputFormat" and thereby removing
one button) is included - commented out with an explanations of how to test.
The tampering won't work when "outputFormat" is included in the JWT, but
by modifying STANDALONE_MOVE_INTO_JWT to prevent that - the tamper will take
effect can be observed. The test establishes that the protection mechanism is
working.